backend/requirements.txt
fastapi
uvicorn
python-jose
psycopg2-binary
azure-storage-blob
python-dotenv
